I believe that I have most of the code synced with onnv_73, keeping in mind not much changed with ZFS between onnv_72 and onnv_73. Look at the output from a "hg log -r onnv_72:onnv_73" in the OpenSolaris code for details.
There is some weirdness with references to "sfsp->f_mntonname" where onnv_73 wants to use an "entry." something or other but this doesn't work.

I've tried as much as possible to put all possible differences between the OpenSolaris source and the Apple code with #ifndef __APPLE__ and such, however as with the noted difference, this wasn't obvious in the code so there are surely places where Apple changed code but didn't make note. In time we should be able to figure this out.