Created an PackageMager pmdoc that should allow a package to be build when the "Release" build is built. Documentation files are sparse, but they are easy to edit and will be auto-updated in the installer package when modified and rebuilt.
Added script which runs fix-nofs when that component is installed with the installer and the OS is at least 10.6.

Rebuild the "Release" configuration, based it upon "Debug" so it might still have some debug junk in the final product, needs refinement I think.

Playing around with the ZFS icon provided by Raoul <tangles@mac.com> on the zfs-macos list. It still needs refinements (the perspective seems off to me) but it's a start and better than anything I could do in any number of years of time.